WIP: 2025/03/06.
The README table doesn't need the `|` at the end.
Simplify this to avoid having to add a lot of spaces by removing the ending `|`.

Add a temporary version called `kevin` to make it easy to explicitly force using this code.

Add `spring-boot-autoconfigure` because the `AnyNestedCondition` needs to be added to support condition on OR.
I am open to better solutions.

Add the Simple condition that is designed to operate without a true OKAPI, such as a static GitHub pages website.

The simple design only performs a fetch of the exact JSON.
Additional logic or designs might be required to handle the `latest` concept.